% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EntrepreneurCompaniesWorkedFor.R
\name{EntrepreneurCompaniesWorkedFor}
\alias{EntrepreneurCompaniesWorkedFor}
\title{Companies that YC and TC entrepreneurs have worked for}
\usage{
EntrepreneurCompaniesWorkedFor(out.file, min.num = 9)
}
\arguments{
\item{out.file}{}

\item{min.num}{is the minimum number of companies that have to show in the dataset}
}
\value{
Writes to a table
}
\description{
Companies that YC and TC entrepreneurs have worked for
}

