% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StanfordCSgradsPhD.R
\name{StanfordCSgradsPhD}
\alias{StanfordCSgradsPhD}
\title{Returns a database of Stanford CS PhDs with details on career history}
\usage{
StanfordCSgradsPhD()
}
\description{
Returns a database of Stanford CS PhDs with details on career history
}

