% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Q3_stanford_cs_phd_grads_career_histories.R
\name{StanfordCSPhDGradsCareerHistories}
\alias{StanfordCSPhDGradsCareerHistories}
\title{Returns a plot of fraction of career trajectories.}
\usage{
StanfordCSPhDGradsCareerHistories()
}
\value{
ggplot object
}
\description{
Returns a ggplot2 plot of Stanford CS PhD choices over time.
}

