% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stanford_cs_phd_grads_occ_choice_per_year.R
\name{StanfordCSPhDGradsOccChoicePerYear}
\alias{StanfordCSPhDGradsOccChoicePerYear}
\title{Returns a plot of fraction of Stanford CS PhD grads choosing various career options over time.}
\usage{
StanfordCSPhDGradsOccChoicePerYear()
}
\value{
String
}
\description{
Returns a ggplot2 plot of Stanford CS PhD grad occupational choices
}

