% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StanfordCSPhDGradsPerYear.R
\name{StanfordCSPhDGradsPerYear}
\alias{StanfordCSPhDGradsPerYear}
\title{Returns a plot of the count of Stanford CS PhD graduates per year.}
\usage{
StanfordCSPhDGradsPerYear()
}
\value{
ggplot object
}
\description{
Returns a ggplot2 plot of Stanford CS PhD graduates per year.
}

