% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Q4_stanford_cs_phd_grads_timeline.R
\name{StanfordCSPhDGradsTimeline}
\alias{StanfordCSPhDGradsTimeline}
\title{Returns a plot of timeline for a single graduate}
\usage{
StanfordCSPhDGradsTimeline(id, grads_only = FALSE)
}
\arguments{
\item{id}{graduate ID}

\item{grads_only}{logical, if only positions after graduation should be plotted}
}
\value{
ggplot object
}
\description{
Returns a ggplot2 plot of Stanford CS PhD timeline
}

