% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Q5_stanford_cs_phd_career_choices.R
\name{StanfordCSPhDGradsCareerChoices}
\alias{StanfordCSPhDGradsCareerChoices}
\title{Individual choices about careers}
\usage{
StanfordCSPhDGradsCareerChoices()
}
\value{
dataset
}
\description{
Create a data set of individual choices about careers at different points in time
}

